% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_morphalou3}
\alias{import_morphalou3}
\title{Importer le lexique Morphalou 3.1}
\usage{
import_morphalou3(...)
}
\arguments{
\item{"..."}{parametres supplemantaires pour readr}
}
\description{
Retourne un tibble contenant le lexique
}
\examples{
\dontrun{
# pas de barre de progression
import_morphalou3(progress = F)
# test sur dix premières lignes
import_morphalou3(n_max 10)
}

}
