% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{licence_Morphalou}
\alias{licence_Morphalou}
\title{Lire la licence de Morphalou}
\usage{
licence_Morphalou()
}
\description{
ouvrir le fichier licence de Morphalou (LGPLLR)
}
\examples{
\dontrun{
licence_Morphalou()
}

}
