% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{lisez_moi}
\alias{lisez_moi}
\title{Ouvrir la doc}
\usage{
lisez_moi()
}
\description{
Retourne un tibble contenant le lexique
}
\details{
N.B. : Les données statistiques et les conjugaisons de la documentation ont été omises dans le package
à cause de leur volume (200 Mo)
}
\examples{
\dontrun{
lisez_moi()
}

}
